# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nook library seat booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeskSession {
  id?: number;
  studentName: string;
  checkinAt: string;
  sessionExpiresAt: string;
  /** @nullable */
  awayAt?: string | null;
  /** @nullable */
  awayExpiresAt?: string | null;
  /** @nullable */
  promptSentAt?: string | null;
  /** @nullable */
  promptRespondedAt?: string | null;
  status: string;
  /** @nullable */
  releasedAt?: string | null;
}

export interface DeskInfo {
  id: string;
  floor: number;
  zone: string;
  zoneName: string;
  amenities: string[];
  status: string;
  session?: DeskSession;
}

export interface LibrarianDeskInfo {
  id: string;
  floor: number;
  zone: string;
  zoneName: string;
  amenities: string[];
  status: string;
  session?: DeskSession;
}

export interface DeskCheckin {
  /** @minLength 1 */
  studentName: string;
}

export interface FloorStats {
  floor: number;
  total: number;
  occupied: number;
  away: number;
  free: number;
}

export interface HourlyStats {
  hour: number;
  sessions: number;
}

export interface ZoneStats {
  zone: string;
  zoneName: string;
  floor: number;
  total: number;
  occupied: number;
  away: number;
}

export interface AnalyticsData {
  activeNow: number;
  freeNow: number;
  awayNow: number;
  totalSessionsToday: number;
  abandonedToday: number;
  avgSessionMinutes: number;
  byFloor: FloorStats[];
  byHour: HourlyStats[];
  byZone: ZoneStats[];
}

export interface DemoSeedResult {
  seeded: number;
  cleared: number;
}

export type ListDesksParams = {
floor?: number;
};