# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nook library seat booking API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: desks
    description: Desk operations
  - name: librarian
    description: Librarian dashboard operations
  - name: analytics
    description: Analytics operations
  - name: demo
    description: Demo data operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /desks:
    get:
      operationId: listDesks
      tags: [desks]
      summary: List all desks with current status
      parameters:
        - name: floor
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of desks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeskInfo"

  /desks/{deskId}:
    get:
      operationId: getDeskById
      tags: [desks]
      summary: Get a single desk
      parameters:
        - name: deskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Desk details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeskInfo"
        "404":
          description: Desk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /desks/{deskId}/checkin:
    post:
      operationId: checkinDesk
      tags: [desks]
      summary: Check in to a desk (QR scan)
      parameters:
        - name: deskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeskCheckin"
      responses:
        "200":
          description: Checked in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeskInfo"
        "400":
          description: Desk not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Desk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /desks/{deskId}/away:
    post:
      operationId: markAway
      tags: [desks]
      summary: Mark desk as away (pauses session for 20 minutes)
      parameters:
        - name: deskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked away
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeskInfo"
        "400":
          description: No active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Desk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /desks/{deskId}/respond:
    post:
      operationId: respondPrompt
      tags: [desks]
      summary: Respond to Still Here prompt — extends session by 2 hours
      parameters:
        - name: deskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session extended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeskInfo"
        "400":
          description: No active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Desk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /desks/{deskId}/release:
    post:
      operationId: releaseDesk
      tags: [desks]
      summary: Release a desk
      parameters:
        - name: deskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Desk released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeskInfo"
        "400":
          description: No active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Desk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /librarian/desks:
    get:
      operationId: listLibrarianDesks
      tags: [librarian]
      summary: List all desks with full session details (librarian only)
      responses:
        "200":
          description: All desks with sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibrarianDeskInfo"

  /librarian/desks/{deskId}/reset:
    post:
      operationId: resetDesk
      tags: [librarian]
      summary: Force-release a desk
      parameters:
        - name: deskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Desk reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeskInfo"
        "404":
          description: Desk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /librarian/desks/{deskId}/alert:
    post:
      operationId: alertDesk
      tags: [librarian]
      summary: Send Still Here alert to a student
      parameters:
        - name: deskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeskInfo"
        "400":
          description: No active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Desk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Get live occupancy analytics
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsData"

  /demo/seed:
    post:
      operationId: seedDemo
      tags: [demo]
      summary: Seed realistic demo sessions for prototype demonstration
      responses:
        "200":
          description: Demo data seeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoSeedResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    DeskSession:
      type: object
      required: [studentName, checkinAt, sessionExpiresAt, status]
      properties:
        id:
          type: integer
        studentName:
          type: string
        checkinAt:
          type: string
          format: date-time
        sessionExpiresAt:
          type: string
          format: date-time
        awayAt:
          type: ["string", "null"]
          format: date-time
        awayExpiresAt:
          type: ["string", "null"]
          format: date-time
        promptSentAt:
          type: ["string", "null"]
          format: date-time
        promptRespondedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
        releasedAt:
          type: ["string", "null"]
          format: date-time

    DeskInfo:
      type: object
      required: [id, floor, zone, zoneName, amenities, status]
      properties:
        id:
          type: string
        floor:
          type: integer
        zone:
          type: string
        zoneName:
          type: string
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
        session:
          $ref: "#/components/schemas/DeskSession"

    LibrarianDeskInfo:
      type: object
      required: [id, floor, zone, zoneName, amenities, status]
      properties:
        id:
          type: string
        floor:
          type: integer
        zone:
          type: string
        zoneName:
          type: string
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
        session:
          $ref: "#/components/schemas/DeskSession"

    DeskCheckin:
      type: object
      required: [studentName]
      properties:
        studentName:
          type: string
          minLength: 1

    FloorStats:
      type: object
      required: [floor, total, occupied, away, free]
      properties:
        floor:
          type: integer
        total:
          type: integer
        occupied:
          type: integer
        away:
          type: integer
        free:
          type: integer

    HourlyStats:
      type: object
      required: [hour, sessions]
      properties:
        hour:
          type: integer
        sessions:
          type: integer

    ZoneStats:
      type: object
      required: [zone, zoneName, floor, total, occupied, away]
      properties:
        zone:
          type: string
        zoneName:
          type: string
        floor:
          type: integer
        total:
          type: integer
        occupied:
          type: integer
        away:
          type: integer

    AnalyticsData:
      type: object
      required: [activeNow, freeNow, awayNow, totalSessionsToday, abandonedToday, avgSessionMinutes, byFloor, byHour, byZone]
      properties:
        activeNow:
          type: integer
        freeNow:
          type: integer
        awayNow:
          type: integer
        totalSessionsToday:
          type: integer
        abandonedToday:
          type: integer
        avgSessionMinutes:
          type: number
        byFloor:
          type: array
          items:
            $ref: "#/components/schemas/FloorStats"
        byHour:
          type: array
          items:
            $ref: "#/components/schemas/HourlyStats"
        byZone:
          type: array
          items:
            $ref: "#/components/schemas/ZoneStats"

    DemoSeedResult:
      type: object
      required: [seeded, cleared]
      properties:
        seeded:
          type: integer
        cleared:
          type: integer