# lib/api-zod/src/generated/types/analyticsData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nook library seat booking API
 * OpenAPI spec version: 0.1.0
 */
import type { FloorStats } from './floorStats';
import type { HourlyStats } from './hourlyStats';
import type { ZoneStats } from './zoneStats';

export interface AnalyticsData {
  activeNow: number;
  freeNow: number;
  awayNow: number;
  totalSessionsToday: number;
  abandonedToday: number;
  avgSessionMinutes: number;
  byFloor: FloorStats[];
  byHour: HourlyStats[];
  byZone: ZoneStats[];
}