# lib/api-zod/src/generated/types/deskInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nook library seat booking API
 * OpenAPI spec version: 0.1.0
 */
import type { DeskSession } from './deskSession';

export interface DeskInfo {
  id: string;
  floor: number;
  zone: string;
  zoneName: string;
  amenities: string[];
  status: string;
  session?: DeskSession;
}