# lib/api-zod/src/generated/types/deskSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nook library seat booking API
 * OpenAPI spec version: 0.1.0
 */

export interface DeskSession {
  id?: number;
  studentName: string;
  checkinAt: Date;
  sessionExpiresAt: Date;
  /** @nullable */
  awayAt?: Date | null;
  /** @nullable */
  awayExpiresAt?: Date | null;
  /** @nullable */
  promptSentAt?: Date | null;
  /** @nullable */
  promptRespondedAt?: Date | null;
  status: string;
  /** @nullable */
  releasedAt?: Date | null;
}