# lib/api-zod/src/generated/types/floorStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nook library seat booking API
 * OpenAPI spec version: 0.1.0
 */

export interface FloorStats {
  floor: number;
  total: number;
  occupied: number;
  away: number;
  free: number;
}